Display all rankings on one screen. (#50)
* Update ranking xml.

* Display all rankings on one screen.

* Make ranking font smaller to not cut off.

* Make predicted rank an int.

+++ app/src/main/java/com/example/viewer_2020/data/DatabaseReference.kt
@@ -85,7 +85,7 @@ class DatabaseReference {
     data class CalculatedPredictedTeam (
         var team_number: Int,
         var predicted_rps: Double,
-        var predicted_rank: Double,
+        var predicted_rank: Int,
         var current_rps: Int,
         var current_rank: Int,
         var current_avg_rps: Float
